+++ src/components/chat/main.tsx
@@ -1,15 +1,13 @@
 'use client';
 
 import { useChat, Message } from 'ai/react';
-import Image from 'next/image';
 import React from 'react';
 
 import { getChatMessages } from '@/services/chatService';
 
-import Logo from '../../../public/media/policy-wonk.png';
-import { ChatMessage } from '../chatMessage';
-
-import Ask from './ask';
+import ChatBox from './chatBox';
+import ChatHeader from './chatHeader';
+import { ChatMessage } from './chatMessage';
 import DefaultQuestions from './defaultQuestions';
 
 const MainContent: React.FC = () => {
@@ -31,26 +29,15 @@ const MainContent: React.FC = () => {
   };
 
   return (
-    <div className='main-content d-flex flex-column'>
-      <Image
-        className='img-fluid policy-png'
-        src={Logo}
-        alt='Aggie Gold Robot cartoon'
-      />
-      <h2 className='main-title'>Policy Wonk</h2>
-      <h3 className='sub-title'>Your UC Policy Expert</h3>
-      <p className='intro-text'>
-        Meet Policywonk, your personal guide to navigating all the ins and outs
-        of UC policies...
-      </p>
+    <div>
+      {messages?.length === 0 && !isLoading && <ChatHeader />}
+      {messages?.length === 0 && isLoading && <p>Loading...</p>}
       {messages
         .filter((m) => m.role === 'assistant' || m.role === 'user')
         .map((m: Message) => (
           <div key={m.id}>
             <strong>{`${m.role}: `}</strong>
             <ChatMessage message={m} />
-            <br />
-            <br />
           </div>
         ))}
 
@@ -59,14 +46,10 @@ const MainContent: React.FC = () => {
         allowSend={!isLoading}
       />
 
-      <Ask onQuestionSubmitted={onQuestionSubmitted} allowSend={!isLoading} />
-
-      <p className='disclaimer-text small mt-2'>
-        Disclaimer: The information provided by Policywonk is for general
-        informational purposes only and should not be considered legal or
-        professional advice. Always consult with the appropriate experts and
-        refer to official policies for accurate and up-to-date information.
-      </p>
+      <ChatBox
+        onQuestionSubmitted={onQuestionSubmitted}
+        allowSend={!isLoading}
+      />
     </div>
   );
 };
